Stats have a 64bits, not 32bits, uptime number

+++ src/amule.cpp
@@ -1406,8 +1406,8 @@ void CamuleApp::OnTCPTimer(AMULE_TIMER_EVENT_CLASS& WXUNUSED(evt))
 void CamuleApp::OnCoreTimer(AMULE_TIMER_EVENT_CLASS& WXUNUSED(evt))
 {
 	// Former TimerProc section
-	static uint32	msPrev1, msPrev5, msPrevSave, msPrevHist, msPrevOS, msPrevKnownMet;
-	uint32 msCur = theStats::GetUptimeMillis();
+	static uint64	msPrev1, msPrev5, msPrevSave, msPrevHist, msPrevOS, msPrevKnownMet;
+	uint64 msCur = theStats::GetUptimeMillis();
 
 	if (!IsRunning()) {
 		return;
